Add files via upload
No metas. Note that the pressure plate button has two objects: the large block embedded in the group that will display either a red X or green O on the large display to indicate whether it's active, and the small pressure plate that blocks get slid onto that is its own object for simple animation of the button.